Section 20 file upload and download

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require('../models/product')
 const { validationResult } = require('express-validator')
+const fileHelper = require('../util/file')
 
 exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
@@ -14,9 +15,25 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title
-  const imageUrl = req.body.imageUrl
+  const image = req.file
   const price = req.body.price
   const description = req.body.description
+
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description
+      },
+      validationErrors: [],
+      errorMessage: 'Attached File is not an image'
+    })
+  }
 
   const errors = validationResult(req)
   if (!errors.isEmpty()) {
@@ -26,13 +43,18 @@ exports.postAddProduct = (req, res, next) => {
       path: '/admin/add-product',
       editing: false,
       hasError: true,
-      product: { title: title, imageUrl: imageUrl, price: price, description: description },
+      product: {
+        title: title,
+        price: price,
+        description: description
+      },
       validationErrors: errors.array(),
       errorMessage: errors.array()[0].msg
     })
   }
+  const imageUrl = image.path
+
   const product = new Product({
-    // _id: mongoose.Types.ObjectId('600c44123185c21ab8944914'),
     title: title,
     price: price,
     description: description,
@@ -44,17 +66,6 @@ exports.postAddProduct = (req, res, next) => {
       res.redirect('/admin/products')
     })
     .catch(err => {
-      // return res.status(500).render('admin/edit-product', {
-      //   pageTitle: 'Add Product',
-      //   path: '/admin/add-product',
-      //   editing: false,
-      //   hasError: true,
-      //   product: { title: title, imageUrl: imageUrl, price: price, description: description },
-      //   validationErrors: [],
-      //   errorMessage: 'Database operation failed please try again'
-      // })
-      // res.redirect('/500')
-
       const error = new Error(err)
       error.httpStatusCode = 500
       return next(error)
@@ -97,7 +108,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId
   const updatedTitle = req.body.title
   const updatedPrice = req.body.price
-  const updatedImageUrl = req.body.imageUrl
+  const image = req.file
   const updatedDesc = req.body.description
 
   const errors = validationResult(req)
@@ -108,7 +119,12 @@ exports.postEditProduct = (req, res, next) => {
       path: '/admin/add-product',
       editing: true,
       hasError: true,
-      product: { title: updatedTitle, imageUrl: updatedImageUrl, price: updatedPrice, description: updatedDesc, _id: prodId },
+      product: {
+        title: updatedTitle,
+        price: updatedPrice,
+        description: updatedDesc,
+        _id: prodId
+      },
       validationErrors: errors.array(),
       errorMessage: errors.array()[0].msg
     })
@@ -122,7 +138,11 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle
       product.price = updatedPrice
       product.description = updatedDesc
-      product.imageUrl = updatedImageUrl
+
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl)
+        product.imageUrl = image.path
+      }
       product.save()
         .then(result => {
           console.log('RESULT: ', result)
@@ -157,8 +177,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId
-
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product Not Found'))
+      }
+      fileHelper.deleteFile(product.imageUrl)
+      return Product.deleteOne({ _id: prodId, userId: req.user._id })
+    })
     .then(() => {
       res.redirect('/admin/products')
     })
